requirements.txt
fastapi>=0.115.12
filelock>=3.13.1
python-dotenv>=1.1.0
python-multipart>=0.0.20
pydantic-settings>=2.9.1
requests>=2.32.0
syft-core>=0.2.3
uvicorn>=0.34.2
git+https://github.com/OpenMined/rds.git@rasswanth/add-enclave-enc
